+++ pkg/domain/password/model.go
@@ -13,21 +13,25 @@ type domain struct {
 }
 
 // NewDomain ...
-func NewDomain(chars []rune) Domain {
+func NewDomain(chars string) Domain {
+	return domain{[]rune(chars)}
+}
+
+func newDomain(chars []rune) Domain {
 	return domain{chars}
 }
 
 func (d domain) Len() uint                 { return uint(len(d.chars)) }
 func (d domain) AsSlice() []rune           { return d.chars }
 func (d domain) Rune(idx uint) rune        { return d.chars[idx] }
-func (d domain) MergeWith(o Domain) Domain { return NewDomain(append(d.chars, o.AsSlice()...)) }
+func (d domain) MergeWith(o Domain) Domain { return newDomain(append(d.chars, o.AsSlice()...)) }
 
 // Most common domains
 var (
-	LowerCaseLetters = NewDomain([]rune("abcdefghijklmnopqrstuvwxyz"))
-	UpperCaseLetters = NewDomain([]rune("ABCDEFGHIJKLMNOPQRSTUVWXYZ"))
-	Numbers          = NewDomain([]rune("0123456789"))
-	ASCIISpecials    = NewDomain([]rune(`!"#$%&'()*+,-./:;<=>?@[\]^_{|}~`))
+	LowerCaseLetters = NewDomain("abcdefghijklmnopqrstuvwxyz")
+	UpperCaseLetters = NewDomain("ABCDEFGHIJKLMNOPQRSTUVWXYZ")
+	Numbers          = NewDomain("0123456789")
+	ASCIISpecials    = NewDomain(`!"#$%&'()*+,-./:;<=>?@[\]^_{|}~`)
 
 	Letters           = LowerCaseLetters.MergeWith(UpperCaseLetters)
 	LettersAndNumbers = Letters.MergeWith(Numbers)