+++ CHANGELOG.md
@@ -17,7 +17,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Add and remove users from groups
 - Export one or all units
 - Import units
-- Set user password, available hash formats : MD5, SMD5, SHA, SSHA, SHA256, SSHA256, SHA384, SSHA384, SHA512, SSHA512
+- Set user password, available hash formats : MD5, SMD5, SHA, SHA1, SSHA, SSHA1, SHA224, SSHA224, SHA256, SSHA256, SHA384, SSHA384, SHA512, SSHA512, SHA3-224, SSHA3-224, SHA3-256, SSHA3-256, SHA3-384, SSHA3-384, SHA3-512, SSHA3-512
 - Password sent by mail to user when changed
 - JSON, YAML, Table outputs for all commands
 - JSON input for all commands